Rename histogram and warningtimer to be pep8 compliant
